backend/team3/requirements.txt
django
djangorestframework
django-rest-auth
django-cors-headers
drf-nested-serializer
coverage
django-cleanup
